Update factor.js
- separate into todoModules.js and projectModules.js
