--- src/createTodo.js
@@ -1,5 +1,0 @@
-import { todoModule } from './factory';
-
-export default (title, description, priority, dueDate, projectId) => {
-  todoModule.createTodo(title, description, priority, dueDate, projectId);
-};
