+++ src/renderDomElements/renderModal.js
@@ -1,5 +1,6 @@
 import { domNodeCreator, chainAppend } from '../domNodeCreator';
-import { projectModule, todoModule } from '../factory';
+import { projectModule } from '../projectModule';
+import { todoModule } from '../todoModule';
 
 export default (index) => {
   const body = document.querySelector('body');
